Refactor SelectTimeZone (Components)
