+++ src/app/components/atoms/country-list/SelectTimeZone.module.css
@@ -1,21 +1,23 @@
-.item-country-container {
+& .item-country-container {
     display: flex;
     width: 100%;
+    min-height: 48px;
+    align-items: center;
     padding: 8px 12px;
     border: 1.5px solid var(--color-darkgray-500);
     background-color: transparent;
-    color: white;
+    color: var(--color-whywhite-500);
     cursor: pointer;
 
     &:active {
-        border: 1.5px solid var(--color-purple-500);
+        border: 1.5px solid var(--color-purple-500)
     }
 
     &:focus {
         outline: 1.5px solid var(--color-purple-500)
     }
 
-    .flag-and-country-name {
+    & .flag-and-country-name {
         display: flex;
         width: 100%;
         align-items: center;
@@ -30,5 +32,27 @@
             font-size: var(--font-size-large);
             text-align: left;
         }
+    }
+
+    & .more-countries-button {
+        width: 22px;
+        height: 22px;
+        background-color: transparent;
+        transform: rotate(-0.25turn);
+        transition: transform 0.05s linear;
+    }
+
+    & .more-countries-button.rotated {
+        transform: rotate(0deg);
+    }
+}
+
+& .more-time-zones {
+    display: none;
+    width: 100%;
+
+    &.is-open {
+        display: flex;
+        flex-direction: column;
     }
 }
