+++ src/app/components/atoms/country-list/TimeZones.tsx
@@ -21,6 +21,7 @@ const TimeZones: React.FC<{ timeZone: FlagZone[] }> = ({ timeZone }) => {
               <button
                 key={self.crypto.randomUUID()}
                 className={styles['mtz-item-country-container']}
+                type="button"
               >
                 <div className={styles['mtz-zone-name-container']}>
                   <span>{getZoneName(zoneNames)}</span>
