+++ src/app/components/molecules/country-list/CountryList.module.css
@@ -6,7 +6,7 @@
     height: 100%;
 
     /* *Pensar este overlay */
-    background-color: rgb(138 0 0 / 20%);
+    background-color: rgb(0 0 0 / 45%);
 
     .container-list-with-search {
         display: block;
@@ -21,7 +21,7 @@
 
             & .search-bar {
                 width: 100%;
-                height: 32px;
+                min-height: 48px;
                 padding: 0 8px;
                 border-radius: 10px 0 0;
                 border-top: 1.5px solid var(--color-darkgray-500);
@@ -48,6 +48,9 @@
             }
 
             & .close-modal {
+                display: flex;
+                align-items: center;
+                justify-content: center;
                 border-radius: 0 10px 0 0;
                 border-top: 1.5px solid var(--color-darkgray-500);
                 border-right: 1.5px solid var(--color-darkgray-500);
@@ -66,8 +69,8 @@
         }
 
         .container-list-of-countries {
-            /* Los 32px son la altura de la .search-bar */
-            height: calc(100% - 32px);
+            /* Los 48px son la altura de la .search-bar */
+            height: calc(100% - 48px);
             border: 1.5px solid var(--color-darkgray-500);
             border-radius: 0 0 10px 10px;
             overflow-y: scroll;
